requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
SQLAlchemy==1.4.49
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Flask-Migrate==4.0.5
Flask-Bootstrap4==4.0.2
Python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23